+++ lib/services/w3m_service/w3m_service.dart
@@ -193,6 +193,7 @@ class W3MService with ChangeNotifier implements IW3MService {
     blockchainService.instance = BlockChainService(
       core: _web3App.core,
     );
+
     magicService.instance = MagicService(
       web3app: _web3App,
       enabled: enableEmail,
@@ -330,8 +331,8 @@ class W3MService with ChangeNotifier implements IW3MService {
   Future<void> _setSesionAndChainData(W3MSession w3mSession) async {
     try {
       await _storeSession(w3mSession);
-      final chainId = _currentSelectedChainId ?? w3mSession.chainId;
-      await _setLocalEthChain(chainId, logEvent: false);
+      _currentSelectedChainId = _currentSelectedChainId ?? w3mSession.chainId;
+      await _setLocalEthChain(_currentSelectedChainId!, logEvent: false);
     } catch (e, s) {
       _logger.e(
         '[$runtimeType] _setSesionAndChainData error $e',
@@ -369,7 +370,7 @@ class W3MService with ChangeNotifier implements IW3MService {
     if (_currentSession != null) {
       final chainId = _savedChainId(null);
       if (chainId != null && W3MChainPresets.chains.containsKey(chainId)) {
-        await selectChain(W3MChainPresets.chains[chainId]!, logEvent: false);
+        await _setLocalEthChain(chainId, logEvent: false);
       } else {
         _currentSelectedChainId = chainId;
       }
@@ -457,6 +458,7 @@ class W3MService with ChangeNotifier implements IW3MService {
   }
 
   Future<void> _setLocalEthChain(String chainId, {bool? logEvent}) async {
+    _currentSelectedChainId = chainId;
     final caip2Chain = 'eip155:$_currentSelectedChainId';
     _logger.i('[$runtimeType] set local chain $caip2Chain');
     _currentSelectedChainId = chainId;
@@ -813,6 +815,7 @@ class W3MService with ChangeNotifier implements IW3MService {
       _logger.i('[$runtimeType] Rebuilding session, ending future');
       return;
     } catch (e) {
+      await disconnect();
       await _connectionErrorHandler(e);
     }
   }
@@ -1016,16 +1019,24 @@ class W3MService with ChangeNotifier implements IW3MService {
 
   @override
   Future<List<dynamic>> requestReadContract({
+    required String? topic,
+    required String chainId,
     required DeployedContract deployedContract,
     required String functionName,
+    EthereumAddress? sender,
     List parameters = const [],
   }) async {
     try {
-      // TODO use blockchain-api if possible.
+      if (selectedChain == null) {
+        throw W3MServiceException(
+          'You must select a chain before reading a contract',
+        );
+      }
       return await _web3App.requestReadContract(
         deployedContract: deployedContract,
         functionName: functionName,
         rpcUrl: selectedChain!.rpcUrl,
+        sender: sender,
         parameters: parameters,
       );
     } catch (e) {
@@ -1040,31 +1051,18 @@ class W3MService with ChangeNotifier implements IW3MService {
     required DeployedContract deployedContract,
     required String functionName,
     required Transaction transaction,
+    List<dynamic> parameters = const [],
     String? method,
-    List parameters = const [],
   }) async {
     try {
-      final requestParams = SessionRequestParams(
-        method: MethodsConstants.ethSendTransaction,
-        params: [
-          Transaction.callContract(
-            contract: deployedContract,
-            function: deployedContract.function(functionName),
-            from: transaction.from,
-            value: transaction.value,
-            maxGas: transaction.maxGas,
-            gasPrice: transaction.gasPrice,
-            nonce: transaction.nonce,
-            maxFeePerGas: transaction.maxFeePerGas,
-            maxPriorityFeePerGas: transaction.maxPriorityFeePerGas,
-            parameters: parameters,
-          ).toJson(),
-        ],
-      );
-      return request(
-        topic: topic,
+      return await _web3App.requestWriteContract(
+        topic: topic ?? '',
         chainId: chainId,
-        request: requestParams,
+        deployedContract: deployedContract,
+        functionName: functionName,
+        transaction: transaction,
+        parameters: parameters,
+        method: method,
       );
     } catch (e) {
       rethrow;
@@ -1299,6 +1297,7 @@ class W3MService with ChangeNotifier implements IW3MService {
       _logger.i('[$runtimeType] requestSwitchToChain error $e');
       // if request errors due to user rejection then set the previous chain
       if (_isUserRejectedError(e)) {
+        // fallback to current chain if rejected by user
         await _setLocalEthChain(_currentSelectedChainId!);
         throw JsonRpcError(code: 5002, message: 'User rejected methods.');
       } else {