+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 3.3.4
+
+- Package is no longer maintained. Check readme file
+
+## 3.3.3
+
+- Fix a bug where reading stored chainId failed
+- Added retry to blockchain call when response is empty
+
+## 3.3.1
+
+- Bug fixes and improvements
+
 ## 3.3.0
 
 - One-Click Auth + SIWE implementation